+++ editor/editor-core/src/plugins/hyperlink/__tests__/visual-regression/hyperlink.ts
@@ -140,7 +140,8 @@ describe('Hyperlink', () => {
 
   describe('hyperlink menu with Link Picker Options and ff:lp-link-picker', () => {
     describe('edit link', () => {
-      it('displays correctly when link matches display text', async () => {
+      // FIXME: This test was automatically skipped due to failure on 16/09/2022: https://product-fabric.atlassian.net/browse/ED-15668
+      it.skip('displays correctly when link matches display text', async () => {
         await initEditorWithAdf(page, {
           appearance: Appearance.fullPage,
           adf: hyperlinkAdf,
@@ -156,7 +157,8 @@ describe('Hyperlink', () => {
         await page.click(hyperlinkSelectors.editLinkBtn);
       });
 
-      it('displays correctly when link is different to display text', async () => {
+      // FIXME: This test was automatically skipped due to failure on 16/09/2022: https://product-fabric.atlassian.net/browse/ED-15669
+      it.skip('displays correctly when link is different to display text', async () => {
         await initEditorWithAdf(page, {
           appearance: Appearance.fullPage,
           adf: hyperlinkWithTextAdf,
@@ -174,7 +176,8 @@ describe('Hyperlink', () => {
     });
 
     describe('insert link', () => {
-      it('displays correctly when inside a caption node', async () => {
+      // FIXME: This test was automatically skipped due to failure on 16/09/2022: https://product-fabric.atlassian.net/browse/ED-15670
+      it.skip('displays correctly when inside a caption node', async () => {
         await initEditorWithAdf(page, {
           appearance: Appearance.fullPage,
           adf: mediaWithCaptionAdf,